split downloadr
